Add UniqueCoords transform
