+++ src/transforms.jl
@@ -13,3 +13,4 @@ include("transforms/basic.jl")
 include("transforms/geometric.jl")
 include("transforms/detrend.jl")
 include("transforms/potrace.jl")
+include("transforms/uniquecoords.jl")
